+++ CHANGES.rst
@@ -9,3 +9,4 @@ development
 2018-06-11 0.0.0
 ================
 - Add project skeleton
+- Add PatBase client skeleton with "login" method